add trybuild test
