+++ crates/re_string_interner/tests/trybuild/forbid_borrow.stderr
@@ -1,0 +1,21 @@
+error[E0433]: failed to resolve: use of undeclared crate or module `trybuild`
+  --> tests/trybuild/forbid_borrow.rs:21:13
+   |
+21 |     let t = trybuild::TestCases::new();
+   |             ^^^^^^^^ use of undeclared crate or module `trybuild`
+
+error[E0308]: mismatched types
+  --> tests/trybuild/forbid_borrow.rs:19:37
+   |
+19 |     assert_eq!(Some(42), my_map.get(hello_str).copied());
+   |                                 --- ^^^^^^^^^ expected `&MyString`, found `&str`
+   |                                 |
+   |                                 arguments to this method are incorrect
+   |
+   = note: expected reference `&MyString`
+              found reference `&str`
+note: method defined here
+  --> $RUST/std/src/collections/hash/map.rs
+   |
+   |     pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V>
+   |            ^^^
